- 반복 테스트런 처리 성능 개선 - 반복 테스트런 서비스 분리

+++ src/main/java/com/mindplates/bugcase/biz/testrun/vo/response/TestrunIterationListResponse.java
@@ -41,7 +41,7 @@ public class TestrunIterationListResponse {
     private Boolean autoTestcaseNotAssignedTester;
     private Integer testcaseGroupCount;
     private Integer testcaseCount;
-    private Integer testrunUserCount;
+    private Long testrunUserCount;
 
     public TestrunIterationListResponse(TestrunIterationDTO testrunIteration) {
         this.id = testrunIteration.getId();
